Better login page styles
