+++ packages/ui/src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Switch, BrowserRouter as Router, Route } from 'react-router-dom';
 import { withCookies } from 'react-cookie';
-import Home from './containers/home';
+import Home from './containers/Home';
 import Saved from './containers/Saved';
 import Auth from './auth';
